+++ packages/graphiql-react/src/index.ts
@@ -10,3 +10,4 @@ export * from './components';
 export type { EditorProps, SchemaReference, SlicesWithActions } from './types';
 export type { GraphiQLPlugin } from './stores/plugin';
 export { KEY_MAP, formatShortcutForOS, isMacOs } from './constants';
+export * from './deprecated';