+++ packages/graphiql-react/src/stores/plugin.ts
@@ -1,6 +1,8 @@
-import { ComponentType } from 'react';
+import type { ComponentType } from 'react';
 import type { StateCreator } from 'zustand';
 import type { SlicesWithActions } from '../types';
+import { storageStore } from './storage';
+import { STORAGE_KEY } from '../constants';
 
 export interface GraphiQLPlugin {
   /**